fix: estoque_saldo NUMERIC (Bling devolve saldo fracionario)
O Bling retorna saldo com casas decimais p/ itens por peso (ex.: 651.43).
A coluna era INTEGER -> o UPDATE quebrava ("invalid input syntax for integer").
Troca p/ NUMERIC (ADD + ALTER TYPE idempotentes) e grava o saldo como numeric[].
Arredondar marcaria 0.3 como esgotado por engano.

+++ api/services/bling-sync.ts
@@ -322,7 +322,7 @@ export async function syncEstoqueBatch(budgetMs = 210000): Promise<EstoqueProgre
     if (ourIds.length > 0) {
       await pool.query(
         `UPDATE products p SET estoque_saldo = v.saldo
-         FROM (SELECT unnest($1::int[]) AS oid, unnest($2::int[]) AS saldo) v
+         FROM (SELECT unnest($1::int[]) AS oid, unnest($2::numeric[]) AS saldo) v
          WHERE p.id = v.oid`,
         [ourIds, saldoVals],
       );
